[Autotest] Simply AudioOutput allowed test to reduce flake.
Currently there is a signficant amount of flake in this test due to the
RMS min/max expected. Rather than reducing this down to a level which
would accomidate all devices in the lab (0.05), this test is being
re-written to check the UI to see if the audio is being displayed as
muted or not.

TEST=Over many devices many iterations.
BUG=None

Change-Id: I5a8c1ace87ff104f17fba93497e6b4de6fc4d692
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1921213
Reviewed-by: Ruben Zakarian <rzakarian@chromium.org>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Commit-Queue: Derek Beckett <dbeckett@chromium.org>
Tested-by: Derek Beckett <dbeckett@chromium.org>