[Autotest] ui_utils.py update to decrease flake
Updated the ui_utils in 2 major ways: Changing the calls to "getDesktop"
each time a call is made rather than always getting it once and forever
using that, and changing all calls to JS promises.
The goal of this is to reduce the odds that a sale "getDesktop" is used
when trying to interact with the UI. This is most noticable when you are
working with "new" desktop elements, such as opening the native settings
page, the file explorer, etc.
Additionally, changing to the js promise helps garuntee a callback,
rather than possible race conditions where an object/information is
returned prior to the ui api finishing its call.

BUG=None
TEST= Every single ui test has been re-run.

Change-Id: If34467bd3d7d15f09b1f5f53b25c629512c3445f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1925564
Reviewed-by: Ruben Zakarian <rzakarian@chromium.org>
Tested-by: Derek Beckett <dbeckett@chromium.org>
Commit-Queue: Derek Beckett <dbeckett@chromium.org>