Autotest: Make platform_Rootdev test genenric
The test is applicable only for /dev/sda storage devices
and this patch will make it generic to applicable for other
storage devices such as nvme, mmc etc. This commit gets the
output of rootdev -s -d and checks the output is having
partition or not.

BUG=None
TEST=Ran on DUT locally

Change-Id: I1d4a982d3181703bd3228961bd8cf8e1f477a637
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1935693
Tested-by: Nicolas Boichat <drinkcat@chromium.org>
Commit-Queue: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Kasaiah Bogineni <kasaiah.bogineni@intel.com>
Reviewed-by: Wai-Hong Tam <waihong@google.com>