[autotest] Test subject alternative name match
This CL tests subjective alternative name (SAN) matching against server
certificates. This is done by adding new server certificate signed by
a new certificate authority certificate. The new server certificate
contains a SAN to be used for matching. This SAN is DNS:example.com and
to test this we once set a correct SAN in shill service properties and
expect the supplicant to accept the server certificate and once we add
an incorrect SAN and expect the supplicant to decline the server
certificate.

BUG=chromium:1030652
TEST=network_WiFi_SimpleConnect.wifi_check1x_PEAP and
     network_WiFi_SimpleConnect.wifi_check1x_TTLS

Cq-Depend: chromium:1897535
Change-Id: I892f7c7549ab11c4073fed6699729b0734c7ab53
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1955404
Tested-by: Omar Morsi <omorsi@google.com>
Commit-Queue: Omar Morsi <omorsi@google.com>
Reviewed-by: Brian Norris <briannorris@chromium.org>