faft-cr50: remove changing the rlz in vpd
The vpd rlz isn't used the same way it was when these tests were first
written. The cr50 scripts now skip setting the board id using
check_device and it doesn't have anything to do with the vpd. We don't
need to change the vpd rlz anymore, so this change removes all of the
support for that.

BUG=none
TEST=run firmware_Cr50BID

Change-Id: I7f8683d6f936b8729d7a4e253ab0d855b6d2f3f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1956204
Tested-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Dana Goyette <dgoyette@chromium.org>
Commit-Queue: Mary Ruthven <mruthven@chromium.org>