firmware_Cr50ConsoleCommands: add i2c monitor for sps boards
We added a MONITOR_I2CS_SDA gpio to detect a wedged i2c bus. This gpio
is enabled on all boards, so we need to add the DIOA1 pin configuration
for sps boards. The MONITOR_I2CS_SDA gpio makes DIOA1 an input. When
cr50 reads the straps, it disables the DIOA1 input. This means DIOA1
isn't always in input on spi boards. It's only an input after deep sleep
resume, because that's the only type of reset that doesn't calculate
straps.

BUG=b:146061787
TEST=run on rammus

Change-Id: I7eda5424ed1e170a935ebc7f5f1a40e6e91ccd10
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1960093
Tested-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Dana Goyette <dgoyette@chromium.org>
Commit-Queue: Mary Ruthven <mruthven@chromium.org>