autotest: Use abstract ChameleonConnection instead ServerProxy
We should rely on ChameleonConnection as it handles chameleon reboot with
a automatic reconnection.
- Change the use of chameleon proxy to ChameleonConnection in
audio_board. The ChameleonConnection will handle the socket error and
try to reconnect, which should help to alleviate the issue that
sometimes DUT loses connection to Chameleon and causes the test to fail.
- Throw ChameleonConnectionError instead of socket.error if the RPC call
still fails after reconnection. This will make the test stop with Error
status instead of Fail caused by unhandled error.

BUG=b:146034535
TEST=Run audio tests and verify they still pass

Change-Id: I07fea8099b77747dad7dfec23256807230d8beab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1961849
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Commit-Queue: En-Shuo Hsu <enshuo@chromium.org>
Tested-by: En-Shuo Hsu <enshuo@chromium.org>