[Autotest] Adding a wait option when refreshing policies
Added a wait option when refreshing policies. When enabled, will wait up
to 1 second for the polices to change after the refresh is called. The
wait will end either when the policies are changed, or the timeout(1
second) has expired. If the timeout expires, the test will simply
continue

BUG=None
TEST=policy_ExtensionPolicy (the failing test)

Change-Id: I2599b953c24beaac7631b023403db7ccd28a58a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1960429
Reviewed-by: Ruben Zakarian <rzakarian@chromium.org>
Tested-by: Derek Beckett <dbeckett@chromium.org>
Commit-Queue: Derek Beckett <dbeckett@chromium.org>