moblab: Add CTS P dev version runners.
The list of executed tests are in the xml file fed as |target_plan|
shared with the internal dev version runner. The xml file, as well
as the zip file of the dev version CTS is updated as needed, each
time we need to resort to a new waiver.

BUG=b:144970045
TEST=cheets_CTS_P.arm.waivers

Change-Id: I0d0f79c124b5e344b71b8723d444d940ca3d9717
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1961858
Tested-by: Kazuhiro Inaba <kinaba@chromium.org>
Auto-Submit: Kazuhiro Inaba <kinaba@chromium.org>
Reviewed-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Yoshiki Iguchi <yoshiki@chromium.org>
Commit-Queue: Kazuhiro Inaba <kinaba@chromium.org>