faft: Rework the xmlrpc server to allow quit and disconnect.
When shutting down, it's nice to have a way to cleanly quit and/or
disconnect from the RPC server, without the framework having to ssh
back in to kill the process.  Disconnecting cleanly helps prevent
the RPC client from hanging on a dead connection, by forcing it to
make a new one instead.

In order to make quitting possible, the rpc server now uses the
autotest standard rpc_server module, so the RpcRouter class was
renamed to fit the XmlRpcDelegate.

This change also makes the RPC server itself terminate old instances,
to avoid encountering "address already in use" errors.  This removes
the need for disconnect() to try running pkill on the dut.

The first tests to call disconnect() in the body are the Consecutive
Reboot tests, since they're simple tests that demonstrate a need to
disconnect for safety.

BUG=None
TEST=Run firmware_FAFTRPC.all and several FAFT levels: faft_lv1, faft_lv5

Change-Id: Ib9ba9852553da99453e2dec04944a68a0af2aa5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1968389
Tested-by: Dana Goyette <dgoyette@chromium.org>
Commit-Queue: Dana Goyette <dgoyette@chromium.org>
Reviewed-by: Greg Edelston <gredelston@google.com>