packet_capturer: avoid new (iw >= v4.14) 'set freq' syntax
This CL started using the new '80MHz' option to 'iw ... set freq':
https://crrev.com/c/1877829

But that's only present on new iw, since v4.14:
https://git.kernel.org/pub/scm/linux/kernel/git/jberg/iw.git/commit/?id=159d5e4263403e2e10716168f672d2cd606bcfa2

We haven't upgraded our Whirlwind APs in a little while, so they don't
have this yet. Use this as a stopgap measure.

This procedure is somewhat ripped off what 'iw' does to compute cf1 for
us.

BUG=chromium:1017533, chromium:1005443
TEST=`network_WiFi_SimpleConnect.wifi_check5VHT80_mixed` (which uses
     80MHz capture) on various APs (note: use a separate router+pcap, or
     else you won't actually hit this code path)

Change-Id: I05792c3a1c5804254eeeefdd94f7b2e4727d3f79
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1967946
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Jared Pauletti <pauletti@google.com>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>