reboot: analyze reason for unexpected reboot.
The primary goal of this change is to figure out if a reboot was done
because of a USB ethernet dongle that died. But while we are at it try
to analyze a few standard reasons that might have lead towards the
unexpected device reboot.

BUG=chromium:1031035
TEST=Ran locally.
     FAIL: Autotest client terminated unexpectedly: DUT rebooted during the test run. (kernel out of memory, kernel panic, process out of memory)

Change-Id: I807dc3c6cd039ec1b9151ff1b742022187af6656
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1966768
Commit-Queue: Ilja H. Friedel <ihf@chromium.org>
Tested-by: Ilja H. Friedel <ihf@chromium.org>
Auto-Submit: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>