[stableversion] [autotest] repair non-fatally.
All tests with a dependency on a servo attempt to repair the servo
in the R79 branch. However, in this branch, the repair path attempts
to reach out to cautotest.

We sever the dependency by wrapping the repair step in a try-catch
so failure does not prevent the test from continuing.

BUG=chromium:1033998
TEST=None
Change-Id: I515059f374b44361b5bf0c1e73b99a4e10e76f06
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1968394
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>
Reviewed-by: Garry Wang <xianuowang@chromium.org>
Tested-by: Gregory Nisbet <gregorynisbet@google.com>