bluetooth: add a new bluetooth_AdapterPowerMeasure test
This patch adds a new test to measure the power consumption
during system suspension.

To run the test, a Raspberry Pi is required to act as the
chameleon host which installs the servo dependency as specified
in Cq-Depend:chromium:1981429

The only supported board is 'kukui' for now. The other boards will
be supported when a more generic method is figured out to specify the
config path for servod.

BUG=b:143862071
TEST=Follow the two steps below
Step 1: connect a servo micro between a DUT, i.e., Kukui, and a
Raspberry Pi.

Step 2: run the autotest as
(cros) test_that --autotest_dir ~/trunk/src/third_party/autotest/files/
       --args "chameleon_host=$CHAMELEON_IP" $DUT_IP
       bluetooth_AdapterPowerMeasure.suspension

       For a stress test of 100 iterations, use the control file:
       bluetooth_AdapterPowerMeasure.suspension_100

Cq-Depend:chromium:1981429, chromium:1980403
Change-Id: Id66f8b4dad911a462096fcada561a82113d78066
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1900950
Reviewed-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Reviewed-by: Yun-Hao Chung <howardchung@google.com>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>