Remove cheets_CTS_N.7.1_r29.arm.CtsDramTestCases from smoke suite.
It still fails on novato boards.

More importantly though, the smoke suite should be for tests that
verify the most basic of functionality: system boots and comes up and
connects to network. CTS does not belong there. If this needs to run in
the CQ, it should be in suite:bvt-cq or suite:bvt-inline.

BUG=chromium:1038699
TEST=test_that --pretend 100.91.3.212 suite:smoke --fast | grep CtsDramTest is empty.

Change-Id: Ia2b2a9ac90232ce6c188b4b65aa6b24a9791566e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1986105
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>
Auto-Submit: Jorge Lucangeli Obes <jorgelo@chromium.org>