bluetooth: add bluetooth_AdapterAUSanity
This patch adds a new Sanity test suite for Bluetooth Audio
tests. The first test is an A2DP test. A binaural audio file
is created with a channel of 440 hz sine wave and the other
channel 20,000 hz sine wave. The DUT plays the audio file
and a Raspberry Pi, acting as a Bluetooth headset, records
the audio stream. The spectrum of the recorded stream is
analyzed and should match 440 hz and 20,000 hz for the test
to pass.

BUG=b:147467915, b:147351296
TEST=Conduct the test
(cros) $ test_that --args "chameleon_host=$CHAMELEON_IP" $DUT_IP \
         bluetooth_AdapterAUSanity

Cq-Depend: chromium:2065813, chromium:2064644
Change-Id: Ib61aa71bb705f25edd008a951033497a6871f704
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2064518
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Reviewed-by: Daniel Winkler <danielwinkler@google.com>