utils: Fix caching of hwmon path
There are 3 problems with the caching of the hwmon paths in
_get_hwmon_paths():
1. There's a typo (hwmon vs hwon).
2. A global is attempted to be updated without declaring.
3. We don't key our cache off the parameter.

Together #1, #2, and #3 and up with a correct implementation of
_get_hwmon_paths() but don't do any caching.

Someone tried to solve just #1 and #2 without #3 prior, but that
results in an incorrect implementation of _get_hwmon_paths().  See
CL:1898275 for the revert.

Let's fix it right.

BUG=chromium:941638
TEST=graphics_WebGLAquarium

Change-Id: I0f844718ddaeb80990e00568a3024fe807dcd5df
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1992409
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>