[Autotest] TPM_Utils - Combining rm -rf calls
The rm -rf calls are split into 7 different calls. Locally this takes ~2
seconds to complete, but looks to be taking ~14 in the lab. It is common
for this util to be called twice on every device test (thus adding 28
seconds to a lab test). This should reduce this 7x (7 calls merged to 1)

TEST=policy_DeviceServer.AllowBluetooth_true
BUG=None

Change-Id: I464503d78ec83f8b7d4df02a566fb9f82c8e85f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1995147
Reviewed-by: Katherine Threlkeld <kathrelkeld@chromium.org>
Commit-Queue: Derek Beckett <dbeckett@chromium.org>
Tested-by: Derek Beckett <dbeckett@chromium.org>
Auto-Submit: Derek Beckett <dbeckett@chromium.org>