firmware_test: report more info after failure to shut down.
If the DUT fails to enter G3 state, there are many possible reasons.
For example, if the device encounters a kernel panic, it will reboot
instead of shutting down.

The function now reports the power state, and if the AP power state
is S0, it gives specific failure messages for DUT still online, and
for DUT unexpectedly rebooted.

BUG=b:134942705
TEST=Run firmware_ConsecutiveBoot, with `shutdown -P` altered to `shutdown -r` to cause failure.

Change-Id: I4282a631a88f1ef7706f58ef98e1059a30fb8d2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1994908
Tested-by: Dana Goyette <dgoyette@chromium.org>
Commit-Queue: Dana Goyette <dgoyette@chromium.org>
Reviewed-by: Matthew Blecker <matthewb@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>