cros/faft: Update gen_test_images.sh
1. Changed "dump_fmap" to "futility dump_fmap" because the former
is now a symlink to the latter, and the symlink is not present
in the lxc containers used by lab servers.
2. Replace call to sign_official_build.sh with call to "futility sign"
because sign_official_build.sh and it's dependencies are not
available in the lxc container used in the lab.
Both replacements are no change in functionality.

BUG=chromium:961293, chromium:1003023
TEST=test_that --board=nocturne <IP> suite:fingerprint

Change-Id: I4ad4df0d7b62da1af08ec7d33e26469420af248d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2006016
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Tested-by: Yicheng Li <yichengli@chromium.org>
Tested-by: Tom Hughes <tomhughes@chromium.org>
Commit-Queue: Tom Hughes <tomhughes@chromium.org>