autotest: fix bootchart missing on NVMe devices.
This change fixes the bug that bootchart is missing in test
BootPerfServerCrosPerf if it runs with a DUT with NVMe storage. Fix the
regular expression in extracting the partition index of "rootdev -s"
output so that values like "/dev/nvme0n1p3" are matched correctly.

BUG=chromium:1043854
TEST=test_that $dut BootPerfServerCrosPerf

Change-Id: Ie6395b13924ec33186b81dc828069b70c3192e4c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2008918
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Tested-by: Chinglin Yu <chinglinyu@chromium.org>
Commit-Queue: Chinglin Yu <chinglinyu@chromium.org>