graphics_HwOverlays: Fix kukui canvas* test cases
kukui/krane etc have a fractional deviceAspectRatio and have a nonzero
rotation (screen.orientation.angle) when set in its 0 UI rotation (i.e.
landscape). This breaks all of our graphics_HwOverlays.canvas* tests.

This CL prerrotates the canvas via a scaleZ() CSS transforms, and fixes
the devicePixelRatio by making the canvas itself of integer dimensions,
while having sub-pixel accuracy (see crbug.com/1042110 for why).

It also adds a CSS left offset to allow for canvases to be entirely
inside the screen, which happens somwhoe in krane. (And this needs the
div with "position:relative" style to wrap up the canvas with
"position:absolute").

TEST=test_that  -b ${BOARD} $IP graphics_HwOverlays.canvas* on krane,
and kohaku
BUG=chromium:1046109, chromium:1046445

Change-Id: I2140ab7b2ac59bf5c7e33a5cc9dfda9cc9d4c00d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2024167
Tested-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>