bluetooth: Add SDP* tests to QuickSanity framework
This CL adds the four SDP tests to QuickSanity. It relies on the
bluetooth_tester peer on chameleon, which is currently working on
raspberry pi only. For this reason, this test will throw a TestNA
exception if run with a non-raspi peer. Likewise, the old SDP tests
running outside QuickSanity will continue to run on non-raspi peers,
throwing TestNA if run on raspi peer.

BUG=b:147515943
TEST=QuickSanity on raspi

Change-Id: Ibb1794e6aa26ddb9c7af65904a6e5da9a6f14012
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2026187
Tested-by: Daniel Winkler <danielwinkler@google.com>
Commit-Queue: Daniel Winkler <danielwinkler@google.com>
Reviewed-by: Yun-Hao Chung <howardchung@google.com>