autotest: fix StorageSuspendStress test for removable devices
At suspend time the kernel re-enumerates USB, which causes problems
with hardware_StorageSuspendStress because when resume finishes, fio
tries to send data to the device which has not been re-scanned yet.

Changed fio_test to check if the device is removable and if so to
change the fio script to continue on all errors.

Added a verify only test after SuspendStress has completed to confirm
there was no corruption.

BUG=chromium:1041586
TEST=Ran hardware_StorageSuspendStress.test and
hardware_StorageSuspendStress.test_other

Change-Id: Iea1c47147d372f066869909c36292ea6beba8bb5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2031489
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Tested-by: Alexis Savery <asavery@chromium.org>
Commit-Queue: Alexis Savery <asavery@chromium.org>