autoupdate_StartOOBEUpdate: Use nebraska.py instead of nano_omaha_devserver.py
This CL changes this autotest to use nebraska.py for the update process.
This is only for the cellular tests.

We are moving away from the use cases of nano_omaha_devserver.py and
eventually will deprecate it.

BUG=chromium:999047
TEST=test_that 100.90.30.183 autoupdate_NonBlockingOOBEUpdate.delta --args="running_at_desk=True job_repo_url=http://100.115.168.196:8082/static/reef-release/R81-12860.0.0/autotest/packages"
TEST=test_that 100.90.30.183 autoupdate_ForcedOOBEUpdate.cellular.full
--args="running_at_desk=True
job_repo_url=http://100.115.168.196:8082/static/reef-release/R81-12860.0.0/autotest/packages"
// For this since I didn't have a cellular enabled device, I commented
out the part that setups the celluar and the test went fine working with nebraska.

Change-Id: I09d0785e7127c7d341ccc30104be676f18647e87
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2031493
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>