Fix broken image links in go/faft doc
For some reason, the image links provided as <img src="..." /> render
properly when viewed in-browser from the filesystem, but not when viewed
through Gitiles. The latter is how go/faft is shown. Hence, go/faft was
not displaying images.

This CL replaces the image links in faft-design-doc.md with the ![]()
syntax used by wifi-basics-codelab.md, which does seem to render
properly in Gitiles:
    https://chromium.googlesource.com/chromiumos/third_party/autotest/+/
    refs/heads/master/docs/wifi-basics-codelab.md

TEST=Open the Base and Patchset Gitiles links of this CL from Gerrit.
Expect to see no images when viewing Base, and yes images for Patchset.
BUG=None

Change-Id: Iefd51088e186a7ed8971e9adced23e5f56864c96
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2041693
Tested-by: Greg Edelston <gredelston@google.com>
Reviewed-by: Kevin Shelton <kmshelton@chromium.org>
Commit-Queue: Greg Edelston <gredelston@google.com>