bluetooth: Reset DUT when object_path fault occurs
Currently we have a bug with QCA where the chip gets into a bad
state and we cannot recover from it. When this occurs in the lab,
it causes many sequential test failures. This CL raises the error
as it should, but also triggers a reboot of the DUT to hopefully
prevent these failures from continuing into subsequent tests.

Judging by the failures seen in the lab, it looks like the first
place where the test will fail is in quick_test_init() where the
bluetooth device facade is created. This is where the error will
be caught.

BUG=b:142276989
TEST=single QuickSanity run
Change-Id: I77e1f525d9ce548738163cfeca76cb1e9eb172ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2042896
Commit-Queue: Daniel Winkler <danielwinkler@google.com>
Tested-by: Daniel Winkler <danielwinkler@google.com>
Reviewed-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>