chrome_cr50: retry eraseflashinfo if the command fails
Check the board id after running eraseflashinfo. Retry the command if
the command contains a failure message or if the board id doesn't get
erased. The command retries up to 10 times by default. The process to
update to the eraseflashinfo image takes a couple of minutes, so it's
important to retry enough times we don't need to do it again.

BUG=b:149088352
TEST=run set board id tests

Change-Id: I9bf0463e6d253a387d7a202edc4f89fbd1cca726
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2044880
Reviewed-by: Dana Goyette <dgoyette@chromium.org>
Commit-Queue: Mary Ruthven <mruthven@chromium.org>
Tested-by: Mary Ruthven <mruthven@chromium.org>