FAFT: Additional checks to fix provisioning when no image is extracted
Added checks for valid EC and BIOS images before attempting to join and
return the extracted path. When no EC image is found we will ignore it
and continue with only programming the BIOS. When no BIOS image is found
we raise an error and provisioning will fail.

BUG=b:147248894
TEST=Ran locally with tarball, tested firmware was correctly flashed to
     DUT. Passed on both kohaku (previously also passed) and arcada
     (previously failed due to no EC image).

Change-Id: I3623ce6ce781fba324b3404d11e22c0f21325dd4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2044979
Tested-by: Brent Peterson <brentpeterson@chromium.org>
Commit-Queue: Brent Peterson <brentpeterson@chromium.org>
Reviewed-by: Mary Ruthven <mruthven@chromium.org>