platform_CompromisedStatefulPartition: clear the TPM before test ends
This change clears the TPM at the end of the test, even if an error is
raised.

Before this change, the system would be in a weird state that
tpm_manager db is gone whereas the TPM remains owned. Not sure if
other daemons/apps are affected.

This change might solve the empty lockout password problem in tests
that soft-clear the TPM. See b/148752436 for details.

BUG=b:148752436
TEST=test_that -b soraka <DUP IP> platform_CompromisedStatefulPartition

Change-Id: Ic73ffeef12de7446cc3554ee5963ad532e5212d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2046683
Tested-by: Wei-Cheng Xiao <garryxiao@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Commit-Queue: Wei-Cheng Xiao <garryxiao@chromium.org>