pubsub_utils_unittest: add explicit chromite import
We're indirectly relying on apiclient in chromite in this module.
utils/unittest_suite.py is currently importing chromite all the
time which is updating sys.path for us.  Add the explicit import
here so we can delete that other code.  This helps reduce the
pollution until we can fix it.

BUG=chromium:1050892
TEST=autotest unittests still pass

Change-Id: I93345d487e73aec44faa22975605c53757b936fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2049084
Reviewed-by: Allen Li <ayatane@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>