graphics_utils: split out uinput usage
The only reason the uinput module is installed in the SDK is so this
module can import it when running unittests.  However, there are no
actual tests of the uinput related code, so we end up installing the
uinput module into the SDK all the time for no reason.

Split all the uinput logic out into a sep module to isolate it.
Now we can drop it from the SDK without breaking the unittest.

BUG=chromium:990593
TEST=CQ passes

Change-Id: Iff780263c602381c01bc3ee912601d38dc652f1f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2049823
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>