FAFT: Improved firmware image extraction during provisioning
This change generates a list of all files within the tarball and then
only extracts the first image candidate found. This ensures the
extraction takes a fixed amount of time when an image is found and
avoids the increased timeout when no image is found (e.g. devices with
non-ChromeOS EC).

BUG=b:149419503
TEST=Ran locally on kohaku, tested firmware was correctly extracted and
     flashed to the DUT. Manually modified script to print files and
     verified the list of files was correct and all image candidates
     were found.

Change-Id: Iaf24c7deaeb0ad14b9046b7544be16fe2afbd20f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2052807
Reviewed-by: Shelley Chen <shchen@chromium.org>
Tested-by: Brent Peterson <brentpeterson@chromium.org>
Commit-Queue: Brent Peterson <brentpeterson@chromium.org>