FAFT: Use FAFT config to check for Chrome EC before extracting EC image
Modified extract_ec_image() function to first check if a board has a
Chrome EC before attempting to extract an EC image from the tarball. If
board does not have a Chrome EC then skip attempting to extract an EC
image and ignore re-programming it. If the board does have a Chrome EC
then attempt to extract the EC image and raise an error if no image was
found.

BUG=b:147248894
TEST=Ran locally with firmware tarball on Kohaku, tested firmware was
     correctly programmed to DUT. Modified script to verify EC image
     extraction was ignored for non-Chrome EC board (sarien).

Change-Id: Ic9767e072e829c7520a36e20e53d828dda9481f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2055787
Reviewed-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Shelley Chen <shchen@chromium.org>
Tested-by: Brent Peterson <brentpeterson@chromium.org>