FAFT: Fixed EC version string extraction from binary image
Provisioning fails for multiple different devices because the script
fails to read a version string from the EC image. Similar to how it's
already being done for the BIOS version we get the EC version string
from the device and then use the board/model prefix to search the EC
image for the matching version string.

BUG=b:147248894
TEST=Ran locally on kohaku, tested version strings were correctly
     extracted and firmware was flashed to the DUT.

Change-Id: Iebf142c61d77e7e193d765241a4c5c4f96955bf2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2056980
Reviewed-by: Mary Ruthven <mruthven@chromium.org>
Tested-by: Brent Peterson <brentpeterson@chromium.org>