Wait for completion of update in autoupdate_CannedOmahaUpdate
We need to wait for the completion or failure of the update before
returning from this code, otherwise we will return soon, the nebraska
is gone, we get the update_engine log's partially, can't look for
failure strings in it anymore.

This is causing autoupdate_CatchBadSignatures test to fail randomly.

BUG=None
TEST=test_that 100.90.30.183 autoupdate_CatchBadSignatures -b reef --no-retry

Change-Id: I92ca073b5194c0a1cb47497fcbe5a535c5101984
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2063208
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Jae Hoon Kim <kimjae@chromium.org>