cr50_test: remove ccd watchdog
Servod always runs with dual-v4 in the lab for some reason. Servod will
die if ccd disconnects for too long. Remove the ccd watchdog, so servo
wont randomly die in the middle of tests. We haven't been paying
attention to ccd state in all tests. It can't really be guaranteed on
all boards. We don't want it to kill servod, so we need to remove the
watchdog until the dual-v4 issue is figured out.

BUG=b:149948314
TEST=none

Change-Id: I4371cfea610809b5d9280c96d457daa8654934d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2067650
Commit-Queue: Mary Ruthven <mruthven@chromium.org>
Tested-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>