faft: firmware_RecoveryButton: Test inserting USB stick after power-on
Currently, this test muxes USB stick to DUT when the DUT is off.
And then we power on the DUT to rec mode and check if the DUT boots
from the USB stick.

This CL adds a new test case that:
(1) The USB stick is muxed to the host when the DUT is off.
(2) We power on the DUT to rec mode.
(3) After a delay, we mux the USB stick to the DUT.
(4) Check if the DUT boots from the USB stick.

This test case tries to mimic the use case that a user inserts the
USB stick after the INSERT screen shows up.

BUG=b:150420142
TEST=pass the test on Kohaku

Change-Id: I0ad9910e5f5fe6bc0457e5d57c9bd96db98a8b84
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2079889
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Tested-by: Philip Chen <philipchen@chromium.org>
Commit-Queue: Philip Chen <philipchen@chromium.org>