Add OFFLOAD_DIR on server and client
Creates a directory in the server-side execution of any test and the
client-side execution of a client-side test. Adds a marker file with
timestamp (UTC) to the 'leaf task'; the client if there is a client,
inside SSP if SSP is in use and no client is, or in the server outside
SSP neither of the others is used.
Surfaces the created directory using the $SYNCHRONOUS_OFFLOAD_DIR envvar
on every platform in use.
Retrieves the directory contents to the drone at the conclusion of the
test.

BUG=chromium:1022986
TEST=untested, but runs without errors locally.

Change-Id: I646cadc80c6ab3149f7af279ca60ee7f7ae0c2f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2071737
Commit-Queue: Jacob Kopczynski <jkop@chromium.org>
Tested-by: Jacob Kopczynski <jkop@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>