wifi_client: refine some TestFail messages
Some functions try to dump the latest state when polling fails, but
passing the exception field of poll_for_condition does not work well.
The evaluation of the expression is not lazy and happens before the
function call. As a result, it only prints the state before polling.
(e.g. state=None in wait_for_connection) Fix the bug so that the
message can be more informative.

This cl should not change any test behavior except TestFail message.

BUG=None
TEST=`test_that $DUT network_WiFi_Prefer5Ghz` to verfiy a caller of
wait_for_bsses. Also, remove configure(ap_config_5G) in the test to
make it fail and check the failure having "Found 1, wanted 2".
TEST=`test_that $DUT network_WiFi_Roam.wifi_roamWPA` to verify a caller
of wait_for_connection. Remove the security config in the second AP in
the control file to make the test fail and check the failure have
"state=idle".

Change-Id: I03efb070fc8e5b7207789076e137764f76a268b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2072542
Tested-by: Yen-lin Lai <yenlinlai@google.com>
Commit-Queue: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>