Use update_url everywhere for checking updates
Currently, the check_for_update() function is getting a server and port
as input and constructs the update URL by itself. But this seems to be
too much confusing as tests who need to call this, and already have a
valid update URL, have to dissect their update URL into pieces and then
feed it to this function which is not ideal and is causing a lot of
issues upgrading these tests. This CLs changes that function to get a
update_url parameter that defines where to get the update. In addition
this replaces the use of NebraskaWrapper.get_port() with its
get_update_url() instead which is more versatile.

BUG=chromium:1055972
TEST=test_that 100.90.30.183 autoupdate_EOL -b reef --no-retry --autotest_dir src/third_party/autotest/files
TEST=test_that 100.90.30.183 autoupdate_CatchBadSignatures -b reef --no-retry --autotest_dir src/third_party/autotest/files

Change-Id: I6ea26223254295758369d7a74d58e8b4678d315c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2076688
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>