utils: Remove sanity check from get_ec_temperature
Some of our perf tools attempted to measure the current temperature as
maxmium of all the temperature readings in the system. This includes
ec, thermal_zone, hwmon. The perf tool itself does not care if any of
those readings are invalid but the final measurement.

This CL drops the sanity check for ec readings and let the caller to
decide if any of the readins are sane.

BUG=chromium:1056749
TEST=test_that DUT graphcis_PerfControl

Change-Id: If51829226a630bb5b8e7a44886f677d1106d706a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2079473
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Commit-Queue: Ilja H. Friedel <ihf@chromium.org>
Tested-by: Ilja H. Friedel <ihf@chromium.org>