Remove the need for spawning a new devserver for AU tests
Since now we can pass critical_update=True as a query string to
devserver/nebraska when doing an update, we don't need to spawn a new
devserver on lab devservers to handle this corner case anymore. So we
need to get rid of this from everywhere and just send it as a query
string to update servers.

Also devservers now support delta_payload artifact. So we don't need to
spawn a new devserver just to manually stage them there.

BUG=chromium:1004487, chromium:793434, chromium:1006305
TEST=test_that chromeos1-row2-rack10-host3.cros autoupdate_ForcedOOBEUpdate.full -b reef --no-retry --autotest_dir /home/ahassani/trunk/src/third_party/autotest/files --args="job_repo_url=http://100.115.219.133:8082/static/reef-release/R82-12953.0.0/autotest/packages"
TEST=test_that chromeos6-row3-rack10-host3 autoupdate_NonBlockingOOBEUpdate.delta -b betty --no-retry --autotest_dir /home/ahassani/trunk/src/third_party/autotest/files --args="job_repo_url=http://100.115.245.201:8082/static/betty-release/R82-12933.0.0/autotest/packages"

Change-Id: Id1cba3c840afad9753968125127664411c79d0dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2054758
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>