FAFT: New firmware test case for reading VPD data through sysfs
This test case verifies that VPD data can be read through sysfs. The
test generates random strings that are written to the RO and RW VPD
sections, the device is rebooted and then the test verifies that the
new VPD data can be read back through sysfs and matches the original
random string. The test case fails if VPD data cannot be read back or
if the data read back does not match the original string written.

BUG=chromium:642860
TEST=Tested locally on kohaku

Cq-Depend: chromium:2090584
Change-Id: I59b0f6bae7fe700a7484dfad4a31db1259e6f3ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2083680
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Mary Ruthven <mruthven@chromium.org>
Tested-by: Brent Peterson <brentpeterson@chromium.org>
Commit-Queue: Brent Peterson <brentpeterson@chromium.org>